[Ashwini] refactor ProfileScreen
