+++ WriteUp/Helpers/PrimaryButton.swift
@@ -26,24 +26,6 @@ class PrimaryButton: UIButton {
     }
 }
 
-//class SignOutButton: UIButton {
-//    let button = UIButton()
-//    init(titleText: String){
-//        super.init(frame: CGRect(x: 0, y: 0, width: frame.width, height: frame.height))
-//        self.backgroundColor     = Constant.SecondaryColor
-//        self.titleLabel?.font    = UIFont(name: "georgia", size: 20)
-//        self.setTitle(titleText, for: .normal)
-//        self.setTitleColor(.white, for: .normal)
-//        self.layer.borderColor   = UIColor.white.cgColor
-//        self.layer.cornerRadius  = 5
-//        self.layer.borderWidth   = 4.0
-//        self.translatesAutoresizingMaskIntoConstraints = false
-//    }
-//    required init?(coder: NSCoder) {
-//        fatalError(Constant.initFatalError)
-//    }
-//}
-
 class SecondaryButton: UIButton {
     let button = UIButton(frame: CGRect.zero)
     init(titleText: String){
